updated readme and moved the file
