+++ python/README.md
@@ -16,24 +16,18 @@
 
 ## read_TBE
 
-This project provides a Python script to load and parse TBE (Tabular Data, Attachments, and Comments) files into Python-native data structures. The script extracts metadata, tabular data, attachments, and comments from the file.
+The read_TBE.py script is designed to parse and extract data from TBE (Tabular Binary Export) files into Python-native data structures. The script extracts metadata, tabular data, attachments, and comments from the file.
 
-### Prerequisites
+# Prerequisites
 
 - Python 3.x installed on your machine.
 - No external dependencies required.
 
-### Files
-
-- load_tbe_file.py: Contains the load_tbe_file function, which reads a TBE file and parses it into Python dictionaries.
-- run_tbe_parser.py: The main script that calls load_tbe_file and prints the parsed data to the console.
-- example.tbe: An example TBE file containing sample data to be parsed.
-
-### Usage
+# Usage
 
 1. Clone the repository
 2. Update the file path (if needed)
-3. Run the script - Navigate to the src folder and run <code>python run_tbe_parser.py</code>
+3. Run the script - Navigate to the src folder and run <code>python3 read_TBE.py</code>
 
 ## strip_header
